+++ src/Sources/Bitly/BitlyApiImporter.php
@@ -6,13 +6,10 @@ namespace Shlinkio\Shlink\Importer\Sources\Bitly;
 
 use JsonException;
 use Psr\Http\Client\ClientExceptionInterface;
-use Psr\Http\Client\ClientInterface;
-use Psr\Http\Message\RequestFactoryInterface;
 use Shlinkio\Shlink\Importer\Exception\ImportException;
+use Shlinkio\Shlink\Importer\Http\InvalidRequestException;
+use Shlinkio\Shlink\Importer\Http\RestApiConsumerInterface;
 use Shlinkio\Shlink\Importer\Model\ImportedShlinkUrl;
-use Shlinkio\Shlink\Importer\Sources\Bitly\BitlyApiException;
-use Shlinkio\Shlink\Importer\Sources\Bitly\BitlyApiParams;
-use Shlinkio\Shlink\Importer\Sources\Bitly\BitlyApiProgressTracker;
 use Shlinkio\Shlink\Importer\Sources\ImportSources;
 use Shlinkio\Shlink\Importer\Strategy\ImporterStrategyInterface;
 use Shlinkio\Shlink\Importer\Util\DateHelpersTrait;
@@ -20,25 +17,20 @@ use Throwable;
 
 use function Functional\filter;
 use function Functional\map;
-use function json_decode;
 use function ltrim;
 use function parse_url;
 use function sprintf;
 use function str_starts_with;
 
-use const JSON_THROW_ON_ERROR;
-
 class BitlyApiImporter implements ImporterStrategyInterface
 {
     use DateHelpersTrait;
 
-    private ClientInterface $httpClient;
-    private RequestFactoryInterface $requestFactory;
+    private RestApiConsumerInterface $apiConsumer;
 
-    public function __construct(ClientInterface $httpClient, RequestFactoryInterface $requestFactory)
+    public function __construct(RestApiConsumerInterface $apiConsumer)
     {
-        $this->httpClient = $httpClient;
-        $this->requestFactory = $requestFactory;
+        $this->apiConsumer = $apiConsumer;
     }
 
     /**
@@ -58,7 +50,7 @@ class BitlyApiImporter implements ImporterStrategyInterface
             foreach ($groups as ['guid' => $groupId]) {
                 // Skip groups until the initial one is found
                 $initialGroupFound = $initialGroupFound || $groupId === $initialGroup;
-                if (!$initialGroupFound) {
+                if (! $initialGroupFound) {
                     continue;
                 }
 
@@ -132,23 +124,14 @@ class BitlyApiImporter implements ImporterStrategyInterface
         BitlyApiProgressTracker $progressTracker
     ): array {
         $url = str_starts_with($url, 'http') ? $url : sprintf('https://api-ssl.bitly.com/v4%s', $url);
-        $request = $this->requestFactory->createRequest('GET', $url)->withHeader(
-            'Authorization',
-            sprintf('Bearer %s', $params->accessToken()),
-        );
-        $resp = $this->httpClient->sendRequest($request);
-        $body = (string) $resp->getBody();
-        $statusCode = $resp->getStatusCode();
 
-        if ($statusCode >= 400) {
+        try {
+            return $this->apiConsumer->callApi($url, ['Authorization' => sprintf('Bearer %s', $params->accessToken())]);
+        } catch (InvalidRequestException $e) {
             throw BitlyApiException::fromInvalidRequest(
-                $url,
-                $statusCode,
-                $body,
+                $e,
                 $progressTracker->generateContinueToken() ?? $params->continueToken(),
             );
         }
-
-        return json_decode($body, true, 512, JSON_THROW_ON_ERROR);
     }
 }