+++ test/Sources/Bitly/BitlyApiExceptionTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace ShlinkioTest\Shlink\Importer\Sources\Bitly;
 
 use PHPUnit\Framework\TestCase;
+use Shlinkio\Shlink\Importer\Http\InvalidRequestException;
 use Shlinkio\Shlink\Importer\Sources\Bitly\BitlyApiException;
 
 class BitlyApiExceptionTest extends TestCase
@@ -15,7 +16,10 @@ class BitlyApiExceptionTest extends TestCase
      */
     public function generatesExpectedMessage(?string $continueToken): void
     {
-        $e = BitlyApiException::fromInvalidRequest('something.com', 500, 'Error body', $continueToken);
+        $e = BitlyApiException::fromInvalidRequest(
+            InvalidRequestException::fromResponseData('something.com', 500, 'Error body'),
+            $continueToken,
+        );
 
         self::assertEquals(
             'Request to Bitly API v4 to URL "something.com" failed with status code "500" and body "Error body"',